seperate components for llm query feature
